Add support for coloring using pygments.
This functionality requires the pygmentize script to be present on the
system in a location pointed to by shell's PATH variable.

The user may choose which renderer to use by setting the *renderer*
parameter to either :colorize (the default) or :pygments.

*render-code-spans* parameter has been added to determine whether the
inline code spans should be colorized according to the language lexing
rules specified by the *render-code-spans-lang* parameter. These two
replace the *colorize-code-spans-as* parameter, because both renderers
can colorize the inline code spans and pytments does a semi-decent
thing even if no lexing rules are speciffied.

Additionally, support for specifying renderer options has been added.
This allows the user to fine-tune the output. Ie.

```c++|linenos=inline

```