Add persistent pygments rendering process
Spawning a pygmentize process every time a code block needs to be
processed takes up a lot of time. It may take up to minutes to process
a small blog. It is far more efficient to spawn one colorizing process
and keep it running. One can then send code and receive colorized HTML
via standard IO. This is the approach that this patch enables while
keeping the old behavior as a default for the sake of backwards
compatibility.

Fixes #29
Fixes #30