feat: add support for syntax highlighting via Chroma
Chroma[1] is a syntax highlighting tool that is the engine behind
syntax highlighting in Hugo[2]. This commit adds support via the `:chroma`
renderer symbol and also adds a new `*chroma-style*` string variable to change the
appearance of the rendered code block.

The benefits of Chroma is that it's _very_ fast and it's distributed
as a static compiled binary meaning it should be easier to use in various environments.

[1]: https://github.com/alecthomas/chroma
[2]: https://gohugo.io/