reference link with unknown label should probably print the label
commonmark spec definition of full reference link includes
"a link label that matches a link reference definition elsewhere"
so arguably it isn't a reference link in that case and should print
the original text. Even without that argument, it is probably more
useful to print it than not (for debugging if nothing else).